requirements.txt
streamlit
google-generativeai
PyPDF2
faiss-cpu
langchain
